add a file to clear a lil bit the code and add defines to arrow
