+++ include/terminal.h
@@ -41,6 +41,10 @@
 # define ENTER 10
 # define CTRL_D 4
 # define CTRL_C 3
+# define L_ARROW 'D'
+# define R_ARROW 'C'
+# define U_ARROW 'A'
+# define D_ARROW 'B'
 
 // ########################################################
 // #						TERMINAL					  #
