+++ src/memory/memory_alloc.c
@@ -72,7 +72,7 @@ static int	alloc_term(t_minishell *minishell)
 		return (1);
 	}
 	minishell->term->size = 0;
-	minishell->term->cols = 0;
+	minishell->term->cols = ft_strlen(TERMINAL_PROMPT);
 	minishell->term->rows = 0;
 	return (0);
 }
